requirements.txt
python-dotenv==1.0.1
langchain==0.3.3
langchain-openai==0.2.2
langchain_community==0.3.2
Flask==3.0.3
unstructured==0.16.0
faiss-cpu==1.9.0
# sentence-transformers